# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Gateway API
  version: 0.1.0
  description: >-
    MVP public data-plane API. It intentionally supports a documented subset of
    OpenAI-compatible chat completions. Unsupported fields that change semantics
    or billing are rejected instead of silently ignored.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /admin/v1/tenants/{tenantId}/projects/{projectId}/virtual-keys:
    post:
      tags: [Virtual Keys]
      operationId: createVirtualKey
      summary: Create a virtual data-plane credential
      description: The full credential is returned by this operation exactly once.
      security:
        - adminOidc: []
      parameters:
        - $ref: '#/components/parameters/TenantIdPath'
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/AdminActor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVirtualKeyRequest'
      responses:
        '201':
          description: Credential created; the credential field cannot be retrieved later
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedVirtualKey'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
  /admin/v1/tenants/{tenantId}/projects/{projectId}/virtual-keys/{virtualKeyId}:
    get:
      tags: [Virtual Keys]
      operationId: getVirtualKey
      summary: Get safe virtual key metadata
      description: This response never contains the credential secret or its digest.
      security:
        - adminOidc: []
      parameters:
        - $ref: '#/components/parameters/TenantIdPath'
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/VirtualKeyIdPath'
      responses:
        '200':
          description: Safe credential metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualKeyMetadata'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    virtualKey:
      type: http
      scheme: bearer
      bearerFormat: agw_live_...
    adminOidc:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT
  parameters:
...
    TenantIdPath:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VirtualKeyIdPath:
      name: virtualKeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AdminActor:
      name: X-Admin-Actor
      in: header
      required: true
      description: Audited actor resolved by the trusted admin authentication layer.
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
  schemas:
...
    CreateVirtualKeyRequest:
      type: object
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [live, test]
          default: live
        expires_at:
          type: [string, 'null']
          format: date-time
        allowed_models:
          type: [array, 'null']
          maxItems: 256
          description: Null inherits project policy; an empty array denies every model.
          items:
            type: string
            minLength: 1
            maxLength: 128
        limits:
          oneOf:
            - $ref: '#/components/schemas/VirtualKeyLimits'
            - type: 'null'
...
    VirtualKeyLimits:
      type: object
      additionalProperties: false
      properties:
        rpm:
          type: integer
          minimum: 1
        tpm:
          type: integer
          minimum: 1
        concurrency:
          type: integer
          minimum: 1
    VirtualKeyMetadata:
      type: object
      additionalProperties: false
      required:
        - id
        - tenant_id
        - project_id
        - prefix
        - status
        - effective_status
        - version
        - created_at
        - created_by
        - updated_at
        - updated_by
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        prefix:
          type: string
          pattern: '^agw_(live|test)_[a-z0-9]{8,32}$'
          description: Non-secret support and lookup prefix.
        status:
          type: string
          enum: [active, rotating, revoked]
        effective_status:
          type: string
          enum: [active, rotating, revoked, expired, rotation_grace_elapsed]
        expires_at:
          type: [string, 'null']
          format: date-time
        allowed_models:
          type: [array, 'null']
          items:
            type: string
        limits:
          oneOf:
            - $ref: '#/components/schemas/VirtualKeyLimits'
            - type: 'null'
        rotated_from_id:
          type: [string, 'null']
          format: uuid
        rotation_grace_expires_at:
          type: [string, 'null']
          format: date-time
        revoked_at:
          type: [string, 'null']
          format: date-time
        revoked_by:
          type: [string, 'null']
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
    IssuedVirtualKey:
      type: object
      additionalProperties: false
      required: [metadata, credential]
      properties:
        metadata:
          $ref: '#/components/schemas/VirtualKeyMetadata'
        credential:
          type: string
          writeOnly: true
          description: Returned only by create/rotate; never retrievable, logged, or persisted.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, type, request_id, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            type:
              type: string
            param:
              type: [string, 'null']
            request_id:
              type: string
            retryable:
              type: boolean
            retry_after_ms:
              type: [integer, 'null']
              minimum: 0
...
  responses:
    InvalidRequest:
      description: Invalid or unsupported request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource was not found within the authorized tenant scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RequestTooLarge:
      description: Request body, header, or content part exceeds configured limits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
security:
  - virtualKey: []